+++ src/Formio.js
@@ -14,6 +14,7 @@ import _defaults from 'lodash/defaults';
 import { eachComponent } from './utils/utils';
 import jwtDecode from 'jwt-decode';
 import './polyfills';
+import CDN from './CDN';
 
 const { fetch, Headers } = fetchPonyfill({
   Promise: NativePromise
@@ -500,6 +501,15 @@ class Formio {
 
   accessInfo() {
     return Formio.accessInfo(this);
+  }
+
+  /**
+   * Sets OAuth Logout URL.
+   *
+   * @return {*}
+   */
+   oauthLogoutURI(uri, options) {
+    return Formio.oauthLogoutURI(uri, Object.assign({ formio: this }, this.options, options));
   }
 
   /**
@@ -1154,6 +1164,14 @@ class Formio {
     }
   }
 
+  static oauthLogoutURI(uri, options) {
+    options = (typeof options === 'string') ? { namespace: options } : options || {};
+    const logoutURIName = `${options.namespace || Formio.namespace || 'formio'}LogoutAuthUrl`;
+    Formio.tokens[logoutURIName];
+    localStorage.setItem(logoutURIName, uri);
+    return Formio.tokens[logoutURIName];
+  }
+
   static setUser(user, opts = {}) {
     const userName = `${opts.namespace || Formio.namespace || 'formio'}User`;
     const storage = localStorage.getItem('useSessionToken') ? sessionStorage : localStorage;
@@ -1349,8 +1367,6 @@ class Formio {
         options
       });
     }
-
-    authUrl = `${Formio.baseUrl}/current`;
     this.currentUserResolved = false;
     return Formio.makeRequest(formio, 'currentUser', authUrl, 'GET', null, options)
       .then((response) => {
@@ -1651,6 +1667,10 @@ Formio.Providers = Providers;
 Formio.version = '---VERSION---';
 Formio.pathType = '';
 Formio.events = new EventEmitter();
+Formio.cdn = new CDN();
+if ((Formio.version || '').includes('rc')) {
+  Formio.cdn.setBaseUrl('https://cdn.test-form.io');
+}
 
 if (typeof global !== 'undefined') {
   Formio.addToGlobal(global);