+++ src/components/datetime/DateTime.unit.js
@@ -12,7 +12,10 @@ import {
   comp6,
   comp7,
   comp8,
- // comp9
+ // comp9,
+  comp10,
+  comp11,
+  comp12
 } from './fixtures';
 
 describe('DateTime Component', () => {
@@ -42,6 +45,27 @@ describe('DateTime Component', () => {
       });
   });
 
+  it('Should not change manually entered value on blur when time is disabled', (done) => {
+    const form = _.cloneDeep(comp11);
+    const element = document.createElement('div');
+
+    Formio.createForm(element, form).then(form => {
+      const dateTime = form.getComponent('dateTime');
+      const blurEvent = new Event('blur');
+
+      const value = '01-02-2021';
+      const input = dateTime.element.querySelector('.input');
+      input.value = value;
+      input.dispatchEvent(blurEvent);
+
+      setTimeout(() => {
+        assert.equal(input.value, value);
+        document.innerHTML = '';
+        done();
+      }, 600);
+    }).catch(done);
+  });
+
   it('Should allow manual input', (done) => {
     const form = _.cloneDeep(comp3);
     const element = document.createElement('div');
@@ -60,6 +84,28 @@ describe('DateTime Component', () => {
         assert.equal(dateTime.getValue().startsWith(expectedValueStart), true);
         assert.equal(dateTime.dataValue.startsWith(expectedValueStart), true);
 
+        document.innerHTML = '';
+        done();
+      }, 300);
+    }).catch(done);
+  });
+
+  it('Should allow manual input for date with full month format (like MMMM)', (done) => {
+    const form = _.cloneDeep(comp12);
+    const element = document.createElement('div');
+
+    Formio.createForm(element, form).then(form => {
+      const dateTime = form.getComponent('dateTime');
+      const blurEvent = new Event('blur');
+
+      const value = 'April 22';
+      const expectedValue = 'April/22';
+      const input = dateTime.element.querySelector('.input');
+      input.value = value;
+      input.dispatchEvent(blurEvent);
+
+      setTimeout(() => {
+        assert.equal(input.value, expectedValue);
         document.innerHTML = '';
         done();
       }, 300);
@@ -601,7 +647,35 @@ describe('DateTime Component', () => {
     }).catch(done);
   });
 
-  //TOFIX
+  it('Should provide correct values with time after submission', (done) => {
+    const form = _.cloneDeep(comp10);
+    const element = document.createElement('div');
+
+    Formio.createForm(element, form).then(form => {
+      const dateTime = form.getComponent('dateTime');
+      const textField = form.getComponent('textField');
+
+      dateTime.setValue('2022-04-01T14:00:00.000');
+      textField.setValue('2022-04-01T14:00:00.000');
+
+      setTimeout(() => {
+        const submit = form.getComponent('submit');
+        const clickEvent = new Event('click');
+        const submitBtn = submit.refs.button;
+        submitBtn.dispatchEvent(clickEvent);
+
+        setTimeout(() => {
+          const input1 = dateTime.element.querySelector('.input');
+          const input2 = textField.element.querySelector('.input');
+
+          assert.equal(input1.value, '2022-04-01 02:00 PM');
+          assert.equal(input2.value, '2022-04-01 02:00 PM');
+          done();
+        }, 200);
+      }, 200);
+    }).catch(done);
+  });
+
   // it('Should provide correct date in selected timezone after submission', (done) => {
   //   const form = _.cloneDeep(comp9);
   //   const element = document.createElement('div');