+++ src/components/selectboxes/SelectBoxes.js
@@ -205,7 +205,7 @@ export default class SelectBoxesComponent extends RadioComponent {
     const maxCount = this.component.validate.maxSelectedCount;
     const isValid = this.isValid(data, dirty);
 
-    if (maxCount || minCount) {
+    if ((maxCount || minCount) && !this.shouldSkipValidation(data, dirty, rowData)) {
       const count = Object.keys(this.validationValue).reduce((total, key) => {
         if (this.validationValue[key]) {
           total++;